# lab8/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Electroscooter Flask Api",
    "description": "This is a demo project (REST Api) written in Flask. It allows users to manipulate a small database of electro scooters.",
    "version": "1.0.11"
  },
  "tags": [
    {
      "name": "electro-scooters",
      "description": "REST methods for manipulating electro scooters database",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/electro-scooters": {
      "post": {
        "tags": [
          "electro-scooters"
        ],
        "summary": "Add electro scooter into the database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "My scooter"
                  },
                  "battery_level": {
                    "type": "number",
                    "example": 80
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succesful scooter addition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter created successfully"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/electro-scooters/{scooter_id}": {
      "parameters": [
        {
          "name": "scooter_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "description": "The ID of the scooter"
        }
      ],
      "get": {
        "tags": [
          "electro-scooters"
        ],
        "summary": "Get Electro Scooter by ID",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Scooter 1"
                    },
                    "battery_level": {
                      "type": "number",
                      "example": 80
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "electro-scooters"
        ],
        "summary": "Update Electro Scooter by ID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Scooter 2",
                    "required": false
                  },
                  "battery_level": {
                    "type": "number",
                    "example": 70,
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter updated succesfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Scooter not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "Delete-Password",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Password for scooter deletion"
          }
        ],
        "tags": [
          "electro-scooters"
        ],
        "summary": "Delete Electro Scooter by ID",
        "responses": {
          "200": {
            "description": "Succesful delete of the scooter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Incorrect deletion password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Incorrect password"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Electro Scooter not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}